v0.9.6 - Update Checkpoints

+++ src/external/libhydrogen/hydrogen.h
@@ -275,7 +275,7 @@ int hydro_kx_nk_3(hydro_kx_state *state, hydro_kx_session_keypair *kp,
 
 void hydro_pwhash_keygen(uint8_t master_key[hydro_pwhash_MASTERKEYBYTES]);
 
-int hydro_pwhash_deterministic(uint8_t *h, size_t h_len, const char *passwd, int passwd_len,
+int hydro_pwhash_deterministic(uint8_t *h, int h_len, const char *passwd, int passwd_len,
                                const char    ctx[hydro_pwhash_CONTEXTBYTES],
                                const uint8_t master_key[hydro_pwhash_MASTERKEYBYTES],
                                int opslimit, int memlimit, uint8_t threads);
